requirements.txt
google-adk
litellm
fastapi
uvicorn
httpx
pydantic
streamlit
python-dotenv